Narrow indexed constructor lookup internals
