+++ core-relations/src/lib.rs
@@ -36,7 +36,7 @@ pub use pool::{Pool, PoolSet, Pooled};
 pub use query::{
     CachedPlan, QueryBuilder, QueryError, RuleBuilder, RuleId, RuleSet, RuleSetBuilder,
 };
-pub use row_buffer::{TaggedRowBuffer, ValueVec};
+pub use row_buffer::TaggedRowBuffer;
 pub use table::{MergeFn, SortedWritesTable};
 pub use table_spec::{
     ColumnId, Constraint, MutationBuffer, Offset, Rebuilder, Row, Table, TableChange, TableSpec,
